fix: همسان‌سازی کلید مصرف خوراک با رکورد اصلی

+++ src/SheepMonitor.Core/Models/FeedConsumptionItem.cs
@@ -3,7 +3,7 @@ namespace SheepMonitor.Core.Models;
 public sealed class FeedConsumptionItem
 {
     public int Id { get; set; }
-    public int FeedConsumptionRecordId { get; set; }
+    public long FeedConsumptionRecordId { get; set; }
     public string FeedCode { get; set; } = string.Empty;
     public decimal PlannedKg { get; set; }
     public decimal ActualKg { get; set; }
